fix: resolve all mypy linting errors
- Fix capability_loader.py: add type: ignore for setattr on Callable
- Fix untyped function in unified_extractor.py: add proper type annotations
- Fix no-any-return issues: use int() cast for embedding memory, use random.choice() instead of np.random
- Fix FileSystemEventHandler: add type: ignore comment for conditional import
- Fix service_provider.py: change dict to Dict for type consistency, remove redundant cast
- Fix framework __init__.py: use setattr with type: ignore for dynamic module attribute
- Fix async iterator signature: change AsyncIterator to AsyncGenerator
- Fix models.py: add proper type ignore for conditional import

+++ victor/storage/embeddings/service.py
@@ -283,7 +283,7 @@ class EmbeddingService:
         Returns:
             Estimated memory usage in bytes
         """
-        return embedding.size * self._bytes_per_float
+        return int(embedding.size * self._bytes_per_float)
 
     def _evict_cache_for_memory(self, needed_bytes: int) -> int:
         """Evict oldest cache entries to free memory.
